+++ extensions/quick-access/src/components/action-on-files.tsx
@@ -15,7 +15,7 @@ import { alertDialog, refreshNumber } from "../hooks/hooks";
 import { LocalStorageKey } from "../utils/constants";
 import React from "react";
 import { copyFileByPath } from "../utils/applescript-utils";
-import { upRank } from "../search-pinned-folders";
+import { upRank } from "../search-pins";
 import { isDirectory } from "../utils/common-utils";
 import { FolderPage } from "./folder-page";
 
@@ -24,18 +24,18 @@ export function ActionRemoveAllDirectories(props: { setRefresh: React.Dispatch<R
   return (
     <Action
       icon={Icon.ExclamationMark}
-      title={`Unpin All Folders`}
+      title={`Unpin All`}
       shortcut={{ modifiers: ["ctrl", "shift"], key: "x" }}
       onAction={async () => {
         await alertDialog(
           Icon.ExclamationMark,
-          "Unpin All Folders",
-          "Are you sure you  want to unpin all folders?",
+          "Unpin All",
+          "Are you sure you  want to unpin all files and folders?",
           "Unpin All",
           async () => {
             await LocalStorage.setItem(LocalStorageKey.LOCAL_PIN_DIRECTORY, JSON.stringify([]));
             setRefresh(refreshNumber);
-            await showToast(Toast.Style.Success, "Success!", `All folders are unpinned.`);
+            await showToast(Toast.Style.Success, "Success!", `All files and folders are unpinned.`);
           }
         );
       }}